requirements_new.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.65.0
anyio==4.10.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
backports.tarfile==1.2.0
banks==2.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
Brotli==1.1.0
bs4==0.0.2
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
click==8.2.1
cmudict==1.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
courlan==1.3.2
cryptography==45.0.7
cssselect==1.3.0
dataclasses-json==0.6.7
dateparser==1.2.2
debugpy==1.8.16
decorator==5.2.1
deepl==1.22.0
Deprecated==1.2.18
dirtyjson==1.0.8
Distance==0.1.3
distro==1.9.0
docutils==0.22
durationpy==0.10
emoji==1.2.0
executing==2.2.1
fastapi==0.116.1
feedparser==6.0.11
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.44.1
gradio_client==1.12.1
greenlet==3.2.4
griffe==1.13.0
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hangul-jamo==1.0.1
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jamo==0.4.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jpype1==1.6.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jusText==3.0.2
keybert==0.9.0
keyring==25.6.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
kollocate==0.0.2
konlpy==0.6.0
koparadigm==0.10.0
kss==6.0.5
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.3
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.32
langchain-tavily==0.2.11
langchain-teddynote==0.4.4
langchain-text-splitters==0.3.11
langchain-upstage==0.7.3
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
llama-cloud==0.1.37
llama-cloud-services==0.6.63
llama-index-core==0.13.4
llama-index-instrumentation==0.4.0
llama-index-readers-llama-parse==0.5.0
llama-index-workflows==1.3.0
llama-parse==0.6.63
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
nltk==3.9.1
numpy==2.3.2
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.22.1
openai==1.104.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pecab==1.0.8
pillow==11.3.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
pip==25.2
platformdirs==4.4.0
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.2.4
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.2
rank-bm25==0.2.2
readability-lxml==0.8.4.1
readme_renderer==44.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.12.11
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentence-transformers==5.1.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tavily-python==0.7.11
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tld==0.13.1
tokenizers==0.20.3
tomlkit==0.13.3
torch==2.8.0
tornado==6.5.2
tossi==0.3.1
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.3
twine==6.1.0
typer==0.17.3
types-requests==2.32.4.20250809
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
Whoosh==2.7.4
wrapt==1.17.3
xlrd==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0